archive/test_path_optimizer.py
"""
测试路径后处理优化模块

验证风和能耗对路径的局部微调效果
"""

import numpy as np
import matplotlib.pyplot as plt
from core.path_optimizer import (
    compute_segment_cost,
    compute_total_cost,
    optimize_path_with_wind_energy,
    optimize_path
)


def create_test_path():
    """创建一条测试路径（模拟巡检路径）"""
    # 一条带有爬升和转向的路径
    path = [
        (0, 0, 50),
        (20, 10, 52),
        (40, 15, 55),
        (60, 20, 60),   # 爬升段
        (80, 25, 58),
        (100, 30, 55),
        (120, 35, 53),
        (140, 40, 50),
        (160, 45, 48),
        (180, 50, 50),
        (200, 55, 52),
        (220, 60, 55),
    ]
    return path


def visualize_optimization(original_path, optimized_path, wind_vector):
    """可视化优化效果"""
    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(14, 6))

    # 提取坐标
    orig_x = [p[0] for p in original_path]
    orig_y = [p[1] for p in original_path]
    opt_x = [p[0] for p in optimized_path]
    opt_y = [p[1] for p in optimized_path]

    # 左图：路径对比
    ax1.plot(orig_x, orig_y, 'b-o', label='原始路径', linewidth=2, markersize=6)
    ax1.plot(opt_x, opt_y, 'r--s', label='优化路径', linewidth=2, markersize=6)

    # 绘制风向
    wind_norm = wind_vector / (np.linalg.norm(wind_vector) + 1e-6) * 20
    ax1.arrow(100, 10, wind_norm[0], wind_norm[1],
              head_width=5, head_length=5, fc='green', ec='green', linewidth=3)
    ax1.text(100, -10, f'风向\n({wind_vector[0]:.1f}, {wind_vector[1]:.1f})',
             ha='center', fontsize=12, color='green')

    ax1.set_xlabel('X (米)', fontsize=12)
    ax1.set_ylabel('Y (米)', fontsize=12)
    ax1.set_title('路径优化对比（俯视图）', fontsize=14)
    ax1.legend(fontsize=12)
    ax1.grid(True, alpha=0.3)
    ax1.axis('equal')

    # 右图：高度对比
    orig_z = [p[2] for p in original_path]
    opt_z = [p[2] for p in optimized_path]

    ax2.plot(range(len(original_path)), orig_z, 'b-o', label='原始高度', linewidth=2, markersize=6)
    ax2.plot(range(len(optimized_path)), opt_z, 'r--s', label='优化高度', linewidth=2, markersize=6)

    ax2.set_xlabel('路径点索引', fontsize=12)
    ax2.set_ylabel('高度 Z (米)', fontsize=12)
    ax2.set_title('高度变化对比', fontsize=14)
    ax2.legend(fontsize=12)
    ax2.grid(True, alpha=0.3)

    plt.tight_layout()
    plt.savefig('figures/path_optimization_test.png', dpi=150, bbox_inches='tight')
    print("\n[保存] 可视化结果: figures/path_optimization_test.png")


def test_basic_optimization():
    """测试基本优化功能"""
    print("=" * 70)
    print("测试1: 基本路径优化功能")
    print("=" * 70)

    # 创建测试路径
    path = create_test_path()
    print(f"\n[测试路径] {len(path)} 个点")
    print(f"  起点: {path[0]}")
    print(f"  终点: {path[-1]}")

    # 设置风向量（从西向东，较强）
    wind_vector = np.array([5.0, 0.0])  # 东风，5m/s
    print(f"\n[风况] 向量: {wind_vector}, 方向: 东风")

    # 计算原始成本
    original_cost = compute_total_cost(path, wind_vector)
    print(f"\n[原始成本]")
    print(f"  总成本: {original_cost['total_cost']:.2f}")
    print(f"  平均成本: {original_cost['avg_cost']:.2f}")
    print(f"  最大成本: {original_cost['max_cost']:.2f}")

    # 执行优化
    print(f"\n[执行优化]...")
    optimized_path = optimize_path_with_wind_energy(
        path,
        wind_vector,
        max_shift=2.0,  # 最大偏移2米
        wind_alpha=0.5,
        climb_factor=2.0,
        cost_threshold_factor=1.2
    )

    # 计算优化后成本
    optimized_cost = compute_total_cost(optimized_path, wind_vector)
    print(f"\n[优化后成本]")
    print(f"  总成本: {optimized_cost['total_cost']:.2f}")
    print(f"  平均成本: {optimized_cost['avg_cost']:.2f}")
    print(f"  最大成本: {optimized_cost['max_cost']:.2f}")

    # 对比
    improvement = original_cost['total_cost'] - optimized_cost['total_cost']
    improvement_pct = (improvement / original_cost['total_cost']) * 100
    print(f"\n[改善效果]")
    print(f"  成本降低: {improvement:.2f}")
    print(f"  改善比例: {improvement_pct:.1f}%")

    # 计算点偏移
    max_offset = 0
    total_offset = 0
    for i, (p1, p2) in enumerate(zip(path, optimized_path)):
        offset = np.linalg.norm(np.array(p1) - np.array(p2))
        total_offset += offset
        max_offset = max(max_offset, offset)
        if offset > 0.1:
            print(f"  点{i}: 偏移 {offset:.2f}m")

    avg_offset = total_offset / len(path)
    print(f"\n[偏移统计]")
    print(f"  平均偏移: {avg_offset:.2f}m")
    print(f"  最大偏移: {max_offset:.2f}m")

    # 可视化
    visualize_optimization(path, optimized_path, wind_vector)

    return {
        'original_cost': original_cost['total_cost'],
        'optimized_cost': optimized_cost['total_cost'],
        'improvement': improvement,
        'improvement_pct': improvement_pct,
        'avg_offset': avg_offset,
        'max_offset': max_offset
    }


def test_different_wind_directions():
    """测试不同风向下的优化效果"""
    print("\n" + "=" * 70)
    print("测试2: 不同风向下的优化效果")
    print("=" * 70)

    path = create_test_path()
    wind_directions = [
        (0, "东风"),
        (90, "北风"),
        (180, "西风"),
        (270, "南风"),
    ]

    results = []
    for angle, name in wind_directions:
        wind_rad = np.radians(angle)
        wind_vector = np.array([
            5.0 * np.cos(wind_rad),
            5.0 * np.sin(wind_rad)
        ])

        original_cost = compute_total_cost(path, wind_vector)['total_cost']
        optimized_path = optimize_path_with_wind_energy(
            path, wind_vector, max_shift=2.0
        )
        optimized_cost = compute_total_cost(optimized_path, wind_vector)['total_cost']

        improvement = original_cost - optimized_cost
        improvement_pct = (improvement / original_cost) * 100

        results.append({
            'direction': name,
            'angle': angle,
            'original': original_cost,
            'optimized': optimized_cost,
            'improvement': improvement,
            'improvement_pct': improvement_pct
        })

        print(f"\n[{name}]")
        print(f"  原始成本: {original_cost:.2f}")
        print(f"  优化成本: {optimized_cost:.2f}")
        print(f"  改善: {improvement:.2f} ({improvement_pct:.1f}%)")

    return results


def test_wind_strength_impact():
    """测试不同风速的影响"""
    print("\n" + "=" * 70)
    print("测试3: 风速对优化效果的影响")
    print("=" * 70)

    path = create_test_path()
    wind_speeds = [0, 3, 5, 8, 10]

    print("\n风速\t原始成本\t优化成本\t改善\t改善率")
    print("-" * 60)

    results = []
    for speed in wind_speeds:
        wind_vector = np.array([speed, 0.0])  # 东风

        original_cost = compute_total_cost(path, wind_vector)['total_cost']
        optimized_path = optimize_path_with_wind_energy(
            path, wind_vector, max_shift=2.0
        )
        optimized_cost = compute_total_cost(optimized_path, wind_vector)['total_cost']

        improvement = original_cost - optimized_cost
        improvement_pct = (improvement / original_cost) * 100

        print(f"{speed}m/s\t{original_cost:.1f}\t\t{optimized_cost:.1f}\t\t{improvement:.1f}\t{improvement_pct:.1f}%")

        results.append({
            'speed': speed,
            'original': original_cost,
            'optimized': optimized_cost,
            'improvement': improvement,
            'improvement_pct': improvement_pct
        })

    return results


def main():
    """运行所有测试"""
    print("\n" + "=" * 70)
    print("         路径后处理优化模块 - 功能测试")
    print("=" * 70)

    # 测试1: 基本优化功能
    result1 = test_basic_optimization()

    # 测试2: 不同风向
    result2 = test_different_wind_directions()

    # 测试3: 不同风速
    result3 = test_wind_strength_impact()

    print("\n" + "=" * 70)
    print("测试完成！")
    print("=" * 70)
    print(f"\n关键发现:")
    print(f"  [OK] 路径优化功能正常")
    print(f"  [OK] 风向影响优化效果")
    print(f"  [OK] 风速越大，优化收益越明显")
    print(f"  [OK] 偏移控制在{result1['max_offset']:.2f}米以内")


if __name__ == "__main__":
    main()
