requirements.txt
# UAV 电网巡检项目 - Python 依赖清单
# 更新时间: 2026-03-29

# =====================================================
# 核心依赖 (必需)
# =====================================================
numpy>=1.20.0
scipy>=1.7.0
scikit-image>=0.18.0
Pillow>=8.0.0

# =====================================================
# 规划算法依赖 (可选 - 用于高级分组规划)
# =====================================================
# 如果需要使用 Step 9.2/9.3 分组感知规划功能，请取消下面的注释
# scikit-learn>=0.24.0

# =====================================================
# 可视化依赖 (可选 - 用于 3D 可视化和动画)
# =====================================================
# 如果需要生成交互式 HTML 可视化，请取消下面的注释
# plotly>=5.0.0

# =====================================================
# 开发/调试工具 (可选)
# =====================================================
# matplotlib>=3.3.0  # 用于静态图表调试

# =====================================================
# 安装说明
# =====================================================
# 基础安装 (仅核心功能):
#   pip install -r requirements.txt
#   # 注释掉 scikit-learn 和 plotly 相关行
#
# 完整安装 (包含所有功能):
#   pip install -r requirements.txt
#   # 取消所有注释
#
# 快速安装分组规划依赖:
#   pip install scikit-learn
#
# 快速安装可视化依赖:
#   pip install plotly
